requirements.txt
asgiref==3.2.10
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.1
drf-extensions==0.6.0
pytz==2020.1
sqlparse==0.3.1
